Move reader processing to normalize
